% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conferences.R
\name{Conference_presentations}
\alias{Conference_presentations}
\title{Refereed Papers}
\usage{
Conference_presentations()
}
\description{
This function lists the conferences I have been presented at in markdown format.
}

