% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interests.R
\name{Interests_and_hobbies}
\alias{Interests_and_hobbies}
\title{Interests and Hobbies}
\usage{
Interests_and_hobbies()
}
\description{
This function lists my interests and hobbiesin markdown format.
}

