% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/papers.R
\name{Selected_papers_and_conferences}
\alias{Selected_papers_and_conferences}
\title{A selection of peer-review papers and conference presentations}
\usage{
Selected_papers_and_conferences()
}
\description{
This function lists a selection of papers and presentations in markdown format.
}

