% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skills.R
\name{Software_and_programming_languages}
\alias{Software_and_programming_languages}
\title{Software and programming languages}
\usage{
Software_and_programming_languages()
}
\description{
This function lists my experiance with various software products and languages in markdown format.
}

